Module Logs_syslog_mirage.Udp

UDP syslog

Parameters

module C : Mirage_console.S

module CLOCK : Mirage_clock.PCLOCK

module STACK : Tcpip.Stack.V4V6

Signature

val create : C.t -> STACK.t -> hostname:string -> STACK.IP.ipaddr -> ?port:int -> ?truncate:int -> ?facility:Syslog_message.facility -> unit -> Logs.reporter

create c udp ~hostname ip ~port ~truncate () is reporter, which sends log messages to ip, port via UDP. Upon failure, a message is emitted to the console c. Each message can be truncated: truncate defaults to 65535 bytes. The hostname is part of each syslog message. The port defaults to 514. facility is the default syslog facility (see Logs_syslog.message).